fix: refactor module generation ports to clarify ownership
