+++ project-generator/src/main/kotlin/io/github/cdsap/projectgenerator/generator/buildfiles/BuildFilesGeneratorAndroid.kt
@@ -1,6 +1,6 @@
 package io.github.cdsap.projectgenerator.generator.buildfiles
 
-import io.github.cdsap.projectgenerator.writer.BuildFilesGenerator
+import io.github.cdsap.projectgenerator.generator.BuildFilesGenerator
 import io.github.cdsap.projectgenerator.generator.toml.AndroidToml
 import io.github.cdsap.projectgenerator.model.LanguageAttributes
 import io.github.cdsap.projectgenerator.model.ProjectGraph
