+++ project-generator/src/main/kotlin/io/github/cdsap/projectgenerator/generator/buildfiles/BuildFilesGeneratorJvm.kt
@@ -1,6 +1,6 @@
 package io.github.cdsap.projectgenerator.generator.buildfiles
 
-import io.github.cdsap.projectgenerator.writer.BuildFilesGenerator
+import io.github.cdsap.projectgenerator.generator.BuildFilesGenerator
 import io.github.cdsap.projectgenerator.model.LanguageAttributes
 import io.github.cdsap.projectgenerator.model.ProjectGraph
 import io.github.cdsap.projectgenerator.model.TypeProject
