+++ project-generator/src/main/kotlin/io/github/cdsap/projectgenerator/generator/planner/ModuleClassPlannerAndroidLegacy.kt
@@ -2,7 +2,7 @@ package io.github.cdsap.projectgenerator.generator.planner
 
 import io.github.cdsap.projectgenerator.model.*
 import io.github.cdsap.projectgenerator.NameMappings
-import io.github.cdsap.projectgenerator.writer.ModuleClassPlanner
+import io.github.cdsap.projectgenerator.generator.ModuleClassPlanner
 
 /**
  * Plans what classes each module should have and their dependencies
