+++ project-generator/src/main/kotlin/io/github/cdsap/projectgenerator/generator/planner/ModuleClassPlannerJvm.kt
@@ -1,6 +1,6 @@
 package io.github.cdsap.projectgenerator.generator.planner
 
-import io.github.cdsap.projectgenerator.writer.ModuleClassPlanner
+import io.github.cdsap.projectgenerator.generator.ModuleClassPlanner
 import io.github.cdsap.projectgenerator.model.*
 import io.github.cdsap.projectgenerator.NameMappings
