+++ project-generator/src/main/kotlin/io/github/cdsap/projectgenerator/writer/ModulesWriter.kt
@@ -1,6 +1,11 @@
 package io.github.cdsap.projectgenerator.writer
 
+import io.github.cdsap.projectgenerator.generator.BuildFilesGenerator
+import io.github.cdsap.projectgenerator.generator.ClassGenerator
 import io.github.cdsap.projectgenerator.generator.GeneratedModuleLayout
+import io.github.cdsap.projectgenerator.generator.ModuleClassPlanner
+import io.github.cdsap.projectgenerator.generator.ResourceGeneratorA
+import io.github.cdsap.projectgenerator.generator.TestGenerator
 import io.github.cdsap.projectgenerator.model.LanguageAttributes
 import io.github.cdsap.projectgenerator.model.ProjectGraph
 import io.github.cdsap.projectgenerator.model.TypeOfStringResources
@@ -75,41 +80,4 @@ abstract class ModulesWrite<MODULE_DEF, DICT>(
             layout.testKotlinPackageDir().mkdirs()
         }
     }
-}
-
-interface ModuleClassPlanner<MODULE_DEF> {
-    fun planModuleClasses(node: ProjectGraph): MODULE_DEF
-}
-
-interface ClassGenerator<MODULE_DEF, DICT> {
-    fun generate(
-        moduleDefinition: MODULE_DEF,
-        projectName: String,
-        a: MutableMap<String, CopyOnWriteArrayList<DICT>>
-    )
-
-    fun obtainClassesGenerated(
-        moduleDefinition: MODULE_DEF,
-        classesDictionary: MutableMap<String, CopyOnWriteArrayList<DICT>>
-    ): MutableMap<String, CopyOnWriteArrayList<DICT>>
-}
-
-interface TestGenerator<MODULE_DEF, DICT> {
-    fun generate(
-        moduleDefinition: MODULE_DEF,
-        projectName: String,
-        classesDictionary: MutableMap<String, CopyOnWriteArrayList<DICT>>
-    )
-}
-
-interface BuildFilesGenerator {
-    fun generateBuildFiles(node: ProjectGraph, lang: LanguageAttributes, generateUnitTests: Boolean)
-}
-
-interface ResourceGeneratorA<DICT> {
-    fun generate(
-        node: ProjectGraph, lang: LanguageAttributes,
-        typeOfStringResources: TypeOfStringResources,
-        classesDictionary: MutableMap<String, CopyOnWriteArrayList<DICT>>
-    )
 }
